Fixed weapon not being picked up underwater
The issue was solved by commenting out some lines that are supposed to prevent a collision callback issue.
In tests I could not get any such issue to be produced, but to be safe I have only commented out the code to easily revert changes in case it's needed.

To provide evidence I should be in the right: collision flags are changed the moment the weapon is materialized in the world, which happens before the fluid physics are applied as shown by message printing testing.